# Grundgerüst einer Single Page App mit REST-Backend/Backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: TravelblogApp
  description: "__ TODO: Diese Datei durch eine eigene API-Spezifikation ersetzen. __"
  contact:
    name: Huema Yilmaz, Elena Vaeth, Ayse Kocak
  license:
    name: CC-BY-4.0
    url: https://creativecommons.org/licenses/by/4.0/deed.de
  version: "1.0"
servers:
- url: http://localhost:3000
...
paths:
...
  /check/{id}:
      get:
        tags:
          - Wunsch
        summary: Wunsch lesen
        description: Liest einen einzelnen Wunsch vom Server.
        operationId: get-check
        parameters:
          - name: id
            in: path
            required: true
            style: simple
            explode: false
            schema:
              type: string
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Wish'
          "404":
            description: Not Found
      put:
        tags:
          - Wunsch überschreiben
        summary: ""
        description: "Überschreibt einen vorhandenen Wunsch auf dem Server. Bis auf die\
          \ ID können alle Felder geändert werden. Der Versuch, die ID zu ändern, wird\
          \ vom Server stillschweigend ignoriert."
        operationId: update-check
        parameters:
          - name: id
            in: path
            required: true
            style: simple
            explode: false
            schema:
              type: string
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Wish'
          "404":
            description: Not Found
      patch:
        tags:
          - Einzelne Wünsche ändern
        summary: Einzelne Wunschfelder ändern
        description: "Ändert einzelne Felder eines vorhandenen Wunsches. Zulässig sind\
          \ alle Felder bis auf die ID, da diese nachträglich nicht mehr verändert werden\
          \ darf."
        operationId: patch-check
        parameters:
          - name: id
            in: path
            required: true
            style: simple
            explode: false
            schema:
              type: string
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Wish'
          "404":
            description: Not Found
      delete:
        tags:
          - Wunsch
        summary: Wunsch löschen
        description: Löscht einen Wunsch auf dem Server.
        operationId: delete-check
        parameters:
          - name: id
            in: path
            required: true
            style: simple
            explode: false
            schema:
              type: string
        responses:
          "200":
            description: OK
          "404":
            description: Not Found
components:
  schemas:
...
    HATEOAS_Resource:
      title: HATEOAS-Links einer Resource
      type: object
      properties:
        read:
          $ref: '#/components/schemas/HATEOAS_Link'
        update:
          $ref: '#/components/schemas/HATEOAS_Link'
        patch:
          $ref: '#/components/schemas/HATEOAS_Link'
        delete:
          $ref: '#/components/schemas/HATEOAS_Link'
    HATEOAS_Link:
      title: HATEOAS-Link
      type: object
      properties:
        url:
          type: string
        method:
          type: string
        query_params:
          type: array
          items:
            type: string
...
    Wish:
      title: Wunsch
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/HATEOAS_Resource'
        _id:
          type: string
        title:
          type: string
        author:
          type: string
        description:
          type: string
        place:
          type: string
        time:
          type: string
      example:
        _id: "621ff9a936997df5baba746e"
        title: "Meine Reise"
        author: "Elena"
        description: "..."
        place: "Spanien"
        time: "10 Tage"
...
  securitySchemes: {}